+++ jdaviz/configs/default/plugins/plot_options/plot_options.vue
@@ -39,59 +39,91 @@
       :hint="viewer_multiselect ? 'Select viewers to set options simultaneously' : 'Select the viewer to set options.'"
     />
 
-    <v-row v-if="show_viewer_bounds">
-      <v-expansion-panels popout>
+    <v-row>
+      <v-expansion-panels accordion>
         <v-expansion-panel>
           <v-expansion-panel-header v-slot="{ open }">
             <span style="padding: 6px">Viewer bounds</span>
           </v-expansion-panel-header>
           <v-expansion-panel-content class="plugin-expansion-panel-content">
-            <glue-state-sync-wrapper :sync="viewer_x_min_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('viewer_x_min')">
+            <glue-state-sync-wrapper :sync="x_min_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('x_min')">
               <glue-float-field
-                ref="viewer_x_min"
-                label="Viewer X Min"
-                :value.sync="viewer_x_min_value"
+                ref="x_min"
+                label="X Min"
+                :value.sync="x_min_value"
                 type="number"
-                :step="viewer_x_bound_step"
-                :suffix="viewer_x_unit_value"
+                :step="x_bound_step"
+                :suffix="display_units['spectral']"
               />
             </glue-state-sync-wrapper>
-            <glue-state-sync-wrapper :sync="viewer_x_max_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('viewer_x_max')">
+            <glue-state-sync-wrapper :sync="x_max_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('x_max')">
               <glue-float-field
-                ref="viewer_x_max"
-                label="Viewer X Max"
-                :value.sync="viewer_x_max_value"
+                ref="x_max"
+                label="X Max"
+                :value.sync="x_max_value"
                 type="number"
-                :step="viewer_x_bound_step"
-                :suffix="viewer_x_unit_value"
+                :step="x_bound_step"
+                :suffix="display_units['spectral']"
               />
             </glue-state-sync-wrapper>
-            <glue-state-sync-wrapper :sync="viewer_y_min_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('viewer_y_min')">
+            <glue-state-sync-wrapper :sync="y_min_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('y_min')">
               <glue-float-field
-                ref="viewer_y_min"
-                label="Viewer Y Min"
-                :value.sync="viewer_y_min_value"
+                ref="y_min"
+                label="Y Min"
+                :value.sync="y_min_value"
                 type="number"
-                :step="viewer_y_bound_step"
-                :suffix="viewer_y_unit_value"
+                :step="y_bound_step"
+                :suffix="display_units['flux']"
               />
             </glue-state-sync-wrapper>
-            <glue-state-sync-wrapper :sync="viewer_y_max_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('viewer_y_max')">
+            <glue-state-sync-wrapper :sync="y_max_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('y_max')">
               <glue-float-field
-                ref="viewer_y_max"
-                label="Viewer Y Max"
-                :value.sync="viewer_y_max_value"
+                ref="y_max"
+                label="Y Max"
+                :value.sync="y_max_value"
                 type="number"
-                :step="viewer_y_bound_step"
-                :suffix="viewer_y_unit_value"
+                :step="y_bound_step"
+                :suffix="display_units['flux']"
               />
             </glue-state-sync-wrapper>
-            <plugin-action-button
-              :results_isolated_to_plugin="false"
-              @click="reset_viewer_bounds"
-            >
-              Reset viewer bounds
-            </plugin-action-button>
+            <glue-state-sync-wrapper :sync="zoom_center_x_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('zoom_center_x')">
+              <glue-float-field
+                ref="zoom_center_x"
+                label="X Center"
+                :value.sync="zoom_center_x_value"
+                type="number"
+                :step="zoom_step"
+                :suffix="display_units['image'] || 'pix'"
+              />
+            </glue-state-sync-wrapper>
+            <glue-state-sync-wrapper :sync="zoom_center_y_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('zoom_center_y')">
+              <glue-float-field
+                ref="zoom_center_y"
+                label="Y Center"
+                :value.sync="zoom_center_y_value"
+                type="number"
+                :step="zoom_step"
+                :suffix="display_units['image'] || 'pix'"
+              />
+            </glue-state-sync-wrapper>
+            <glue-state-sync-wrapper :sync="zoom_radius_sync" :multiselect="viewer_multiselect" @unmix-state="unmix_state('zoom_radius')">
+              <glue-float-field
+                ref="zoom_radius"
+                label="Zoom-radius"
+                :value.sync="zoom_radius_value"
+                type="number"
+                :step="zoom_step"
+                :suffix="display_units['image'] || 'pix'"
+              />
+            </glue-state-sync-wrapper>
+            <v-row justify="end">
+              <plugin-action-button
+                :results_isolated_to_plugin="false"
+                @click="reset_viewer_bounds"
+              >
+                Reset viewer bounds
+              </plugin-action-button>
+            </v-row>
           </v-expansion-panel-content>
         </v-expansion-panel>
       </v-expansion-panels>