+++ CHANGES.rst
@@ -7,7 +7,7 @@ New Features
 
 - Stretch histogram shows a spinner when the histogram data is updating. [#2644]
 
-- Spectrum viewer bounds can now be set through the Plot Options UI. [#2604]
+- Spectrum and image viewer bounds can now be set through the Plot Options UI. [#2604, #2649]
 
 - Opacity for spatial subsets is now adjustable from within Plot Options. [#2663]